+++ main.go
@@ -2,20 +2,18 @@ package main
 
 import (
 	"fmt"
-	"io"
 	"log"
-	"os"
 	"time"
 
 	"net/http"
 
 	"github.com/danstis/Plex-Sync/config"
+	"github.com/danstis/Plex-Sync/logger"
 	"github.com/danstis/Plex-Sync/models"
 	"github.com/danstis/Plex-Sync/plex"
 	"github.com/danstis/Plex-Sync/webui"
 	"github.com/gorilla/handlers"
 	"github.com/jinzhu/gorm"
-	"gopkg.in/natefinch/lumberjack.v2"
 )
 
 func main() {
@@ -36,10 +34,10 @@ func main() {
 
 	r := webui.NewRouter()
 
-	loggedRouter := handlers.LoggingHandler(createLogger(config.Settings.General.WebserverLogfile), r)
+	loggedRouter := handlers.LoggingHandler(logger.CreateLogger(config.Settings.General.WebserverLogfile), r)
 	go http.ListenAndServe(fmt.Sprintf(":%v", config.Settings.General.WebserverPort), loggedRouter)
 	log.Printf("Started webserver http://localhost:%v", config.Settings.General.WebserverPort)
-	log.SetOutput(createLogger(config.Settings.General.Logfile))
+	log.SetOutput(logger.CreateLogger(config.Settings.General.Logfile))
 
 	for {
 		token := plex.Token()
@@ -55,13 +53,4 @@ func main() {
 		log.Printf("Sleeping for %v...", config.Settings.General.SyncInterval)
 		time.Sleep(config.Settings.General.SyncInterval)
 	}
-}
-
-func createLogger(filename string) io.Writer {
-	return io.MultiWriter(&lumberjack.Logger{
-		Filename:   filename,
-		MaxSize:    config.Settings.General.MaxLogSize, // megabytes
-		MaxBackups: config.Settings.General.MaxLogCount,
-		MaxAge:     config.Settings.General.MaxLogAge, //days
-	}, os.Stdout)
 }